+++ cedar-policy/src/api.rs
@@ -2691,17 +2691,24 @@ impl TemplateResourceConstraint {
 
 /// Unique ids assigned to policies and templates.
 ///
-/// A `PolicyId` can can be constructed using [`PolicyId::from_str`] or by
+/// A [`PolicyId`] can can be constructed using [`PolicyId::from_str`] or by
 /// calling `parse()` on a string. This currently always returns `Ok()`.
 ///
 /// ```
 /// # use cedar_policy::PolicyId;
-/// let id : PolicyId = "my-id".parse().unwrap();
+/// let id = PolicyId::new("my-id");
 /// # assert_eq!(id.as_ref(), "my-id");
 /// ```
 #[repr(transparent)]
 #[derive(Debug, PartialEq, Eq, Clone, Hash, Serialize, Deserialize, RefCast)]
 pub struct PolicyId(ast::PolicyID);
+
+impl PolicyId {
+    /// Construct a [`PolicyId`] from a source string
+    pub fn new(id: impl AsRef<str>) -> Self {
+        Self(ast::PolicyID::from_string(id.as_ref()))
+    }
+}
 
 impl FromStr for PolicyId {
     type Err = ParseErrors;