+++ cedar-policy/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `PolicyId::new()` added to `PolicyId` (#587, resolving #551)
 - `AsRef<str>` implementation for `PolicyId`. (#504, resolving #503)
 - New API `template_links` for `Policy` to retrieve the linked values for a
   template-linked policy. (#515, resolving #489)